+++ changelog.md
@@ -7,8 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [8.1.0 - 2021-03-04]
 ### Added
 - Bridgehead can set their site via rest api
+- Bridgehead can insert new site
+- Feature toggle
 ### Changed
-- Postgres plugin version 42.2.19 for Postgres 13
+- Postgres plugin version 42.2.19
+- Jooq plugin version 3.11.12
+- Parent pom 11.1.1
 ### Fixed
 - Bridgehead can delete their registration via rest api